% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_voting.R
\name{clusterVoting}
\alias{clusterVoting}
\title{Number of clusters}
\usage{
clusterVoting(data, min.k, max.k, algorithm)
}
\arguments{
\item{data}{A dataframe, where columns are features and rows are data points}

\item{min.k}{Minimum number of clusters for which we calculate stabilities}

\item{max.k}{Maximum number of clusters for which we calculate stabilities}

\item{algorithm}{The clustering algorithm to use for the multiple clustering
runs to be measured}
}
\value{
A list the contains: a matrix with stability scores for every k and internal index,
cluster memberships for every k, a dataframe with the k votes for every index,
the frequency barplot of the k votes
}
\description{
Number of clusters
}
\examples{
clusterVoting(toy_genes, 4,14,"sc")
clusterVoting(toy_genes, 2,7,"hc")
clusterVoting(toy_genes, 2,4,"km")
}
