% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feasibilty_analysis.R
\name{feasibilityAnalysis}
\alias{feasibilityAnalysis}
\title{Simulating dataset and calculate stabilities over different number of
clusters}
\usage{
feasibilityAnalysis(classes = 3, samples = 320, features = 400)
}
\arguments{
\item{classes}{The number of classes of samples to be reflected in the
simulated dataset. Also determines the ks to be considered
(classes-2, classes+2)}

\item{samples}{The number of samples in the simulated dataset}

\item{features}{The number of features in the simulated dataset}
}
\value{
An object of class "feasibilityAnalysis" containing the average
stabilities for all number of clusters(k), the average (over all k) and
maximum stabilities observed and the generated dataset
}
\description{
Simulating dataset and calculate stabilities over different number of
clusters
}
\examples{
feasibilityAnalysis(classes = 2, samples = 20, features = 30)

}
