% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{featureSelection}
\alias{featureSelection}
\title{Predictor variable subsampling sets and bootstrapping stability set selection}
\usage{
featureSelection(data, min.k = 2, max.k = 4, step = 5)
}
\arguments{
\item{data}{A dataframe, where columns are features and rows are data points}

\item{min.k}{Minimum number of clusters for which we calculate stabilities}

\item{max.k}{Maximum number of clusters for which we calculate stabilities}

\item{step}{The number for additional features each feature set will contain}
}
\value{
An object of class "featureSelection" containing the dataframe of
average bootstrap stabilities, where rows represent feature sets and columns
number of clusters, the corresponding line plot, the number and the names of
the selected features
}
\description{
Predictor variable subsampling sets and bootstrapping stability set selection
}
\examples{
featureSelection(toy_genes, min.k = 2, max.k = 4, step = 4)

}
