% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_singatures.R
\name{geneSignatures}
\alias{geneSignatures}
\title{Generating the feature/gene signature per cluster}
\usage{
geneSignatures(data, memberships)
}
\arguments{
\item{data}{A dataframe, where columns are features and rows are data points.
The firs column must be named "id" and containing the sample ids}

\item{memberships}{A dataframe with column "id" (same samples ids as above)
and column "membership" containing the cluster membership of each sample.
The memberships must be strings}
}
\value{
An object of class "geneSignature" containing a list of LASSO
(regression analysis) coefficients of each gene and a plot of the highest
30% of coefficients per cluster.
}
\description{
Generating the feature/gene signature per cluster
}
\examples{
geneSignatures(toy_genes, toy_gene_memberships)
}
