% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_agreement_scores.R
\name{get_agreement_scores}
\alias{get_agreement_scores}
\title{Get a dataframe of partition agreement scores for a set of random parameters
clustering runs across different methods}
\usage{
get_agreement_scores(object)
}
\arguments{
\item{object}{An object of class "partitionAgreement"}
}
\value{
A dataframe of partition agreement scores for a set of random
parameters clustering runs across different methods
}
\description{
Get a dataframe of partition agreement scores for a set of random parameters
clustering runs across different methods
}
\examples{
pa.object <- partitionAgreement(toy_genes, algorithm.1 = "spectral",
measure.1 = "rbfdot", algorithm.2 = "kmeans",measure.2 = "Lloyd",
number.of.clusters = 3)
get_agreement_scores(pa.object)
}
