% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_average_feature_k_stabilities.R
\name{get_average_feature_k_stabilities}
\alias{get_average_feature_k_stabilities}
\title{Get a dataframe of average bootstrap stabilities}
\usage{
get_average_feature_k_stabilities(object)
}
\arguments{
\item{object}{An object of class "featureSelection"}
}
\value{
A dataframe of average bootstrap stabilities
}
\description{
Get a dataframe of average bootstrap stabilities
}
\examples{
fs.object <- featureSelection(toy_genes, min.k = 3, max.k = 4, step = 10)
get_average_feature_k_stabilities(fs.object)
}
