% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_average_stabilities_per_k.R
\name{get_average_stabilities_per_k}
\alias{get_average_stabilities_per_k}
\title{Get average stabilities for all numbers of clusters(k)}
\usage{
get_average_stabilities_per_k(object)
}
\arguments{
\item{object}{An object of class "feasibilityAnalysis"}
}
\value{
Average stabilities for all numbers of clusters(k)
}
\description{
Get average stabilities for all numbers of clusters(k)
}
\examples{
fa.object <- feasibilityAnalysis(classes = 4, samples = 50, features = 15)
average.sts.k <- get_average_stabilities_per_k(fa.object)
}
