% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_average_stability.R
\name{get_average_stability}
\alias{get_average_stability}
\title{Get the average stability(over all k)}
\usage{
get_average_stability(object)
}
\arguments{
\item{object}{An object of class "feasibilityAnalysis"}
}
\value{
The average stability(over all k)
}
\description{
Get the average stability(over all k)
}
