% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cluster_memberships_k.R
\name{get_cluster_memberships_k}
\alias{get_cluster_memberships_k}
\title{Get cluster memberships for every k}
\usage{
get_cluster_memberships_k(object)
}
\arguments{
\item{object}{An object of class "clusterVoting"}
}
\value{
Cluster memberships for every k
}
\description{
Get cluster memberships for every k
}
