% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cluster_voting_k_votes_omada.R
\name{get_cluster_voting_k_votes}
\alias{get_cluster_voting_k_votes}
\title{Get k vote frequencies}
\usage{
get_cluster_voting_k_votes(object)
}
\arguments{
\item{object}{An object of class "clusterAnalysis"}
}
\value{
Matrix with k vote frequencies
}
\description{
Get k vote frequencies
}
\examples{
oa.object <- omada(toy_genes, method.upper.k = 6)
get_cluster_voting_k_votes(oa.object)
}
