% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cluster_voting_metric_votes_omada.R
\name{get_cluster_voting_metric_votes}
\alias{get_cluster_voting_metric_votes}
\title{Get a dataframe with the k votes for every index}
\usage{
get_cluster_voting_metric_votes(object)
}
\arguments{
\item{object}{An object of class "clusterAnalysis"}
}
\value{
Dataframe with the k votes for every index
}
\description{
Get a dataframe with the k votes for every index
}
\examples{
oa.object <- omada(toy_genes, method.upper.k = 4)
get_cluster_voting_metric_votes(oa.object)
}
