% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cluster_voting_scores_omada.R
\name{get_cluster_voting_scores}
\alias{get_cluster_voting_scores}
\title{Get a matrix with metric scores for every k and internal index}
\usage{
get_cluster_voting_scores(object)
}
\arguments{
\item{object}{An object of class "clusterAnalysis"}
}
\value{
A matrix with metric scores for every k and internal index
}
\description{
Get a matrix with metric scores for every k and internal index
}
\examples{
oa.object <- omada(toy_genes, method.upper.k = 4)
get_cluster_voting_scores(oa.object)
}
