% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coefficient_dataset.R
\name{get_coefficient_dataset}
\alias{get_coefficient_dataset}
\title{Get a list of LASSO (regression analysis) coefficients of each gene}
\usage{
get_coefficient_dataset(object)
}
\arguments{
\item{object}{An object of class "geneSignature"}
}
\value{
A list of LASSO (regression analysis) coefficients of each gene
}
\description{
Get a list of LASSO (regression analysis) coefficients of each gene
}
