% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_feature_selection_optimal_features_omada.R
\name{get_feature_selection_optimal_features}
\alias{get_feature_selection_optimal_features}
\title{Get the optimal features}
\usage{
get_feature_selection_optimal_features(object)
}
\arguments{
\item{object}{An object of class "clusterAnalysis"}
}
\value{
The list of optimal features
}
\description{
Get the optimal features
}
