% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/get_feature_selection_optimal_number_of_features_omada.R
\name{get_feature_selection_optimal_number_of_features}
\alias{get_feature_selection_optimal_number_of_features}
\title{Get the optimal number of features}
\usage{
get_feature_selection_optimal_number_of_features(object)
}
\arguments{
\item{object}{An object of class "clusterAnalysis"}
}
\value{
The optimal number of features
}
\description{
Get the optimal number of features
}
