% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_feature_selection_scores_omada.R
\name{get_feature_selection_scores}
\alias{get_feature_selection_scores}
\title{Get a dataframe of average bootstrap stabilities}
\usage{
get_feature_selection_scores(object)
}
\arguments{
\item{object}{An object of class "clusterAnalysis"}
}
\value{
A dataframe of average bootstrap stabilities
}
\description{
Get a dataframe of average bootstrap stabilities
}
\examples{
oa.object <- omada(toy_genes, method.upper.k = 6)
get_feature_selection_scores(oa.object)
}
