% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_generated_dataset.R
\name{get_generated_dataset}
\alias{get_generated_dataset}
\title{Get the simulated dataset}
\usage{
get_generated_dataset(object)
}
\arguments{
\item{object}{An object of class "feasibilityAnalysis"}
}
\value{
Simulated dataset
}
\description{
Get the simulated dataset
}
