% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_internal_metric_scores.R
\name{get_internal_metric_scores}
\alias{get_internal_metric_scores}
\title{Get a matrix with metric scores for every k and internal index}
\usage{
get_internal_metric_scores(object)
}
\arguments{
\item{object}{An object of class "clusterVoting"}
}
\value{
A matrix with metric scores for every k and internal index
}
\description{
Get a matrix with metric scores for every k and internal index
}
\examples{
cv.object <- clusterVoting(toy_genes, 4,6,"sc")
get_internal_metric_scores(cv.object)
}
