% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_max_stability.R
\name{get_max_stability}
\alias{get_max_stability}
\title{Get the maximum stability}
\usage{
get_max_stability(object)
}
\arguments{
\item{object}{An object of class "feasibilityAnalysis"}
}
\value{
The maximum stability
}
\description{
Get the maximum stability
}
\examples{
fa.object <- feasibilityAnalysis(classes = 2, samples = 10, features = 15)
maximum.st <- get_max_stability(fa.object)
}
