% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metric_votes_k.R
\name{get_metric_votes_k}
\alias{get_metric_votes_k}
\title{Get a dataframe with the k votes for every index}
\usage{
get_metric_votes_k(object)
}
\arguments{
\item{object}{An object of class "clusterVoting"}
}
\value{
Dataframe with the k votes for every index
}
\description{
Get a dataframe with the k votes for every index
}
\examples{
cv.object <- clusterVoting(toy_genes, 4,8,"sc")
get_metric_votes_k(cv.object)
}
