% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_optimal_features.R
\name{get_optimal_features}
\alias{get_optimal_features}
\title{Get the optimal features}
\usage{
get_optimal_features(object)
}
\arguments{
\item{object}{An object of class "featureSelection"}
}
\value{
The list of optimal features
}
\description{
Get the optimal features
}
