% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_optimal_memberships.R
\name{get_optimal_memberships}
\alias{get_optimal_memberships}
\title{Get a dataframe with the memberships of the samples found in the input data}
\usage{
get_optimal_memberships(object)
}
\arguments{
\item{object}{An object of class "optimalClustering"}
}
\value{
A dataframe with the memberships of the samples found in the
input data
}
\description{
Get a dataframe with the memberships of the samples found in the input data
}
\examples{
oc.object <- optimalClustering(toy_genes, 4, "spectral")
get_optimal_memberships(oc.object)
}
