% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_optimal_number_of_features.R
\name{get_optimal_number_of_features}
\alias{get_optimal_number_of_features}
\title{Get the optimal number of features}
\usage{
get_optimal_number_of_features(object)
}
\arguments{
\item{object}{An object of class "featureSelection"}
}
\value{
The optimal number of features
}
\description{
Get the optimal number of features
}
\examples{
fs.object <- featureSelection(toy_genes, min.k = 3, max.k = 6, step = 10)
get_optimal_number_of_features(fs.object)
}
