% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_optimal_parameter_used.R
\name{get_optimal_parameter_used}
\alias{get_optimal_parameter_used}
\title{Get the optimal parameter used}
\usage{
get_optimal_parameter_used(object)
}
\arguments{
\item{object}{An object of class "optimalClustering"}
}
\value{
The optimal parameter used
}
\description{
Get the optimal parameter used
}
\examples{
oc.object <- optimalClustering(toy_genes, 4, "spectral")
get_optimal_parameter_used(oc.object)
}
