% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_optimal_stability_score.R
\name{get_optimal_stability_score}
\alias{get_optimal_stability_score}
\title{Get the optimal stability score}
\usage{
get_optimal_stability_score(object)
}
\arguments{
\item{object}{An object of class "optimalClustering"}
}
\value{
The optimal stability score
}
\description{
Get the optimal stability score
}
