% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_partition_agreement_scores.R,
%   R/get_partition_agreement_scores_omada.R
\name{get_partition_agreement_scores}
\alias{get_partition_agreement_scores}
\title{Get a dataframe of partition agreement scores for a set of random parameters
clustering runs across different methods}
\usage{
get_partition_agreement_scores(object)

get_partition_agreement_scores(object)
}
\arguments{
\item{object}{An object of class "clusterAnalysis"}
}
\value{
A dataframe of partition agreement scores for a set of random
parameters clustering runs across different methods

A dataframe of partition agreement scores
parameters clustering runs across different methods
}
\description{
Get a dataframe of partition agreement scores for a set of random parameters
clustering runs across different methods

Get a dataframe of partition agreement scores
}
\examples{
ms.object <- clusteringMethodSelection(toy_genes, method.upper.k = 3,
number.of.comparisons = 2)
get_partition_agreement_scores(ms.object)
oa.object <- omada(toy_genes, method.upper.k = 6)
get_partition_agreement_scores(oa.object)
}
