% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_memberships_omada.R
\name{get_sample_memberships}
\alias{get_sample_memberships}
\title{Get a dataframe with the memberships of the samples found in the input data}
\usage{
get_sample_memberships(object)
}
\arguments{
\item{object}{An object of class "clusterAnalysis"}
}
\value{
A dataframe with the memberships of the samples found in the
input data
}
\description{
Get a dataframe with the memberships of the samples found in the input data
}
