% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top30percent_coefficients.R
\name{get_top30percent_coefficients}
\alias{get_top30percent_coefficients}
\title{Plot of the highest 30 percentage of coefficients per cluster}
\usage{
get_top30percent_coefficients(object)
}
\arguments{
\item{object}{An object of class "geneSignature"}
}
\value{
A plot of the highest 30 percentage of coefficients per cluster
}
\description{
Plot of the highest 30 percentage of coefficients per cluster
}
