% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omada.R
\name{omada}
\alias{omada}
\title{A wrapper function that utilizes all tools to produce the optimal sample memberships}
\usage{
omada(data, method.upper.k = 5)
}
\arguments{
\item{data}{A dataframe, where columns are features and rows are data points}

\item{method.upper.k}{The upper limit of clusters, k, to be considered. Must be more than 2}
}
\value{
A list with all the tool results.
}
\description{
A wrapper function that utilizes all tools to produce the optimal sample memberships
}
\examples{
omada(toy_genes, method.upper.k = 5)
omada(toy_genes, method.upper.k = 3)
}
