% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_clustering.R
\name{optimalClustering}
\alias{optimalClustering}
\title{Clustering with the optimal parameters estimated by these tools}
\usage{
optimalClustering(data, clusters, algorithm)
}
\arguments{
\item{data}{A dataframe, where columns are features and rows are data points}

\item{clusters}{Number of clusters to be generated by this clustering}

\item{algorithm}{The clustering algorithm to be used}
}
\value{
A dataframe with the memberships of the samples found in the input
data
}
\description{
Clustering with the optimal parameters estimated by these tools
}
\examples{
optimalClustering(toy_genes, 4,"spectral")
optimalClustering(toy_genes, 2,"kmeans")
}
