% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_average_stabilities.R
\name{plot_average_stabilities}
\alias{plot_average_stabilities}
\title{Plot the average bootstrap stabilities}
\usage{
plot_average_stabilities(object)
}
\arguments{
\item{object}{An object of class "featureSelection"}
}
\value{
Line plot of average bootstrap stabilities
}
\description{
Plot the average bootstrap stabilities
}
\examples{
fs.object <- featureSelection(toy_genes, min.k = 3, max.k = 6, step = 3)
plot_average_stabilities(fs.object)
}
