% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster_voting_omada.R
\name{plot_cluster_voting}
\alias{plot_cluster_voting}
\title{Plot k vote frequencies}
\usage{
plot_cluster_voting(object)
}
\arguments{
\item{object}{An object of class "clusterAnalysis"}
}
\value{
Plot k vote frequencies
}
\description{
Plot k vote frequencies
}
\examples{
oa.object <- omada(toy_genes, method.upper.k = 3)
plot_cluster_voting(oa.object)
}
