% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_feature_selection_omada.R
\name{plot_feature_selection}
\alias{plot_feature_selection}
\title{Plot the average bootstrap stabilities}
\usage{
plot_feature_selection(object)
}
\arguments{
\item{object}{An object of class "clusterAnalysis"}
}
\value{
Line plot of average bootstrap stabilities
}
\description{
Plot the average bootstrap stabilities
}
\examples{
oa.object <- omada(toy_genes, method.upper.k = 4)
plot_feature_selection(oa.object)
}
