% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signature_feature_omada.R
\name{plot_signature_feature}
\alias{plot_signature_feature}
\title{Plot of the highest 30 percentage of coefficients per cluster}
\usage{
plot_signature_feature(object)
}
\arguments{
\item{object}{An object of class "clusterAnalysis"}
}
\value{
A plot of the highest 30 percentage of coefficients per cluster
}
\description{
Plot of the highest 30 percentage of coefficients per cluster
}
