% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_top30percent_coefficients.R
\name{plot_top30percent_coefficients}
\alias{plot_top30percent_coefficients}
\title{Plot of the highest 30 percentage of coefficients per cluster}
\usage{
plot_top30percent_coefficients(object)
}
\arguments{
\item{object}{An object of class "geneSignature"}
}
\value{
A plot of the highest 30 percentage of coefficients per cluster
}
\description{
Plot of the highest 30 percentage of coefficients per cluster
}
\examples{
gs.object <- geneSignatures(toy_genes, toy_gene_memberships)
plot_top30percent_coefficients(gs.object)
}
