% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes-memberships-data.R
\docType{data}
\name{toy_gene_memberships}
\alias{toy_gene_memberships}
\title{Cluster memberships for toy gene data for package examples}
\format{
An object of class `"cross"`; see [qtl::read.cross()].
}
\source{
nope
}
\usage{
data(toy_gene_memberships)
}
\description{
Column "id" represents genes and column "memberships" represents their respective clusters. Rows are samples
}
\examples{
data(toy_gene_memberships)
}
\references{
nothing
}
\keyword{datasets}
