% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes-data.R
\docType{data}
\name{toy_genes}
\alias{toy_genes}
\title{Toy gene data for package examples}
\format{
An object of class `"cross"`; see [qtl::read.cross()].
}
\source{
nope
}
\usage{
data(toy_genes)
}
\description{
Columns are genes and rows are samples
}
\examples{
data(toy_genes)
}
\references{
nothing
}
\keyword{datasets}
