% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_voting.R
\name{clusterVoting}
\alias{clusterVoting}
\title{Estimating number of clusters through internal exhaustive ensemble majority
voting}
\usage{
clusterVoting(data, min.k, max.k, algorithm)
}
\arguments{
\item{data}{A dataframe, where columns are features and rows are data points}

\item{min.k}{Minimum number of clusters for which we calculate stabilities}

\item{max.k}{Maximum number of clusters for which we calculate stabilities}

\item{algorithm}{The clustering algorithm to use for the multiple clustering
runs to be measured}
}
\value{
A list the contains: a matrix with metric scores for every k and
internal index, cluster memberships for every k, a dataframe with the k votes
 for every index, k vote frequencies and the frequency barplot of the k votes
}
\description{
Estimating number of clusters through internal exhaustive ensemble majority
voting
}
\examples{
clusterVoting(toy_genes, 4,14,"sc")
clusterVoting(toy_genes, 2,7,"hc")
clusterVoting(toy_genes, 2,4,"km")
}
