% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{featureSelection}
\alias{featureSelection}
\title{Feature Selection}
\usage{
featureSelection(data, min.k = 2, max.k = 4, step = 5)
}
\arguments{
\item{data}{A dataframe, where columns are features and rows are data points}

\item{min.k}{Minimum number of clusters for which we calculate stabilities}

\item{max.k}{Maximum number of clusters for which we calculate stabilities}

\item{step}{The number for additional features each feature set will contain}
}
\value{
A dataframe of average bootstrap stabilities, where rows represent
feature sets and columns number of clusters and the corresponding plot
}
\description{
Feature Selection
}
\examples{
featureSelection(toy_genes, min.k = 3, max.k = 9, step = 3)
featureSelection(toy_genes, min.k = 2, max.k = 4, step = 4)

}
