% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_partition_agreement.R,
%   R/plot_partition_agreement_omada.R
\name{plot_partition_agreement}
\alias{plot_partition_agreement}
\title{Plot of partition agreement scores}
\usage{
plot_partition_agreement(object)

plot_partition_agreement(object)
}
\arguments{
\item{object}{An object of class "clusterAnalysis"}
}
\value{
Plot of partition agreement scores

Plot of partition agreement scores
}
\description{
Plot of partition agreement scores

Plot of partition agreement scores
}
\examples{
ms.object <- clusteringMethodSelection(toy_genes, method.upper.k = 3,
number.of.comparisons = 2)
plot_partition_agreement(ms.object)
oa.object <- omada(toy_genes, method.upper.k = 4)
plot_partition_agreement(oa.object)
}
